typedef

+++ include/typedef.h
@@ -2,11 +2,10 @@
 brief:常用基本类型定义
 */
 #pragma once
-
-using uint16 = unsigned short;
-using uint32 = unsigned int;
-using uint64 = unsigned long long;
-using int64 = long long;
+using uint16 = uint16_t;
+using uint32 = uint32_t;
+using uint64 = uint64_t;
+using int64 = int64_t;
 using uint8 = unsigned char;
 using int32 = int;
